+++ tests/test_commands.py
@@ -26,8 +26,8 @@ class CommandsTests(TestCase):
         border = manager.get(cca2='AD').borders.get(cca2='FR')
         self.assertTrue(border.cca2, 'FR')
 
-    def test_command_load_and_dump_countries(self):
-        call_command('load_countries', babel=True, verbosity=0)
+    def test_command_load_and_dumpcountries(self):
+        call_command('loadcountries', babel=True, verbosity=0)
 
         self.assertTrue(models.Country.objects.exists())
         self.assertTrue(models.CountryName.objects.exists())
@@ -38,4 +38,4 @@ class CommandsTests(TestCase):
         self.assertTrue(models.Timezone.objects.exists())
         # self.assertTrue(models.Translation.objects.exists())
 
-        call_command('dump_countries', verbosity=0)
+        call_command('dumpcountries', verbosity=0)